Moved filter from epf module to ximgproc
