--- modules/epf/CMakeLists.txt
@@ -1,2 +1,0 @@
-set(the_description "Esge-Preserving filter for images.")
-ocv_define_module(epf opencv_core opencv_imgproc WRAP python java)
