--- modules/epf/README.md
@@ -1,6 +1,0 @@
-Edge-Preserving Filter Module
-===========
-
-Created by Simon Reich
-
-This module allows you to smooth input data. It reduces Gaussian as well as salt & pepper noise. It runs in real-time.
