--- modules/epf/doc/epf.bib
@@ -1,14 +1,0 @@
- @inproceedings{reichwoergoetterdellen2018,
-  author = {Reich, S. and Wörgötter, F. and Dellen, B.},
-  title = {A Real-Time Edge-Preserving Denoising Filter},
-  pages = {85-94},
-  booktitle = {Proceedings of the 13th International Joint Conference on Computer Vision, Imaging and Computer Graphics Theory and Applications (VISIGRAPP): Visapp},
-  year = {2018},
-  volume= {4},
-  location = {Funchal, Madeira (Portugal)},
-  month = {January 27-29},
-  organization = {INSTICC},
-  publisher = {SciTePress},
-  doi = {10.5220/0006509000850094},
-  abstract = {Even in todays world, where augmented reality glasses and 3d sensors become rapidly less expensive and widely more used, the most important sensor remains the 2d RGB camera. Every camera is an optical device and prone to sensor noise, especially in dark environments or environments with extreme high dynamic range. The here introduced filter removes a wide variation of noise, for example Gaussian noise and salt-and-pepper noise, but preserves edges. Due to the highly parallel structure of the method, the implementation on a GPU runs in real-time, allowing us to process standard images within tens of milliseconds. The filter is first tested on 2d image data and based on the Berkeley Image Dataset and Coco Dataset we outperform other standard methods. Afterwards, we show a generalization to arbitrary dimensions using noisy low level sensor data. As a result the filter can be used not only for image enhancement, but also for noise reduction on sensors like acceleremoters, gyroscopes, or GPS-trackers, which are widely used in robotic applications.}
-  }
