--- modules/epf/doc/epf.rst
@@ -1,6 +1,0 @@
-EPF - Edge-Preserving Filter
-==========================
-
-.. highlight:: cpp
-
-This module allows you to smooth image data. It uses a local method to preserve edges while doing so. Please see attached bib file for further details.
