--- modules/epf/samples/epf_demo.cpp
@@ -1,102 +1,0 @@
-/*M///////////////////////////////////////////////////////////////////////////////////////
-//
-//  IMPORTANT: READ BEFORE DOWNLOADING, COPYING, INSTALLING OR USING.
-//
-//  By downloading, copying, installing or using the software you agree to this license.
-//  If you do not agree to this license, do not download, install,
-//  copy or use the software.
-//
-//
-//                           License Agreement
-//                For Open Source Computer Vision Library
-//
-// Copyright (C) 2000-2008, Intel Corporation, all rights reserved.
-// Copyright (C) 2009, Willow Garage Inc., all rights reserved.
-// Third party copyrights are property of their respective owners.
-//
-// Redistribution and use in source and binary forms, with or without modification,
-// are permitted provided that the following conditions are met:
-//
-//   * Redistribution's of source code must retain the above copyright notice,
-//     this list of conditions and the following disclaimer.
-//
-//   * Redistribution's in binary form must reproduce the above copyright notice,
-//     this list of conditions and the following disclaimer in the documentation
-//     and/or other materials provided with the distribution.
-//
-//   * The name of the copyright holders may not be used to endorse or promote products
-//     derived from this software without specific prior written permission.
-//
-// This software is provided by the copyright holders and contributors "as is" and
-// any express or implied warranties, including, but not limited to, the implied
-// warranties of merchantability and fitness for a particular purpose are disclaimed.
-// In no event shall the Intel Corporation or contributors be liable for any direct,
-// indirect, incidental, special, exemplary, or consequential damages
-// (including, but not limited to, procurement of substitute goods or services;
-// loss of use, data, or profits; or business interruption) however caused
-// and on any theory of liability, whether in contract, strict liability,
-// or tort (including negligence or otherwise) arising in any way out of
-// the use of this software, even if advised of the possibility of such damage.
-//
-//M*/
-//################################################################################
-//
-//                    Created by Simon Reich
-//
-//################################################################################
-
-#include "precomp.hpp"
-
-namespace cv
-{
-namespace epf
-#include <iostream>
-#include <opencv2/core.hpp>
-#include <opencv2/core/utils/trace.hpp>
-#include <opencv2/epf.hpp>
-#include <opencv2/highgui.hpp>
-#include <opencv2/imgcodecs.hpp>
-#include <opencv2/opencv.hpp>
-#include <opencv2/ximgproc.hpp>
-#include <string>
-
-using namespace cv;
-
-int main(int argc, char **argv)
-{
-	CV_TRACE_FUNCTION();
-
-	// Help text
-	cv::CommandLineParser parser(
-	    argc, argv,
-	    "{help h ?  | | help message }"
-	    "{@image    | | Path to image file to process }");
-	if (parser.has("help") || !parser.has("@image"))
-	{
-		parser.printMessage();
-		return 0;
-	}
-
-	// Load image from first parameter
-	std::string filename = argv[1];
-	Mat image = imread(filename, 1), res;
-	if (!image.data)
-	{
-		std::cerr << "No image data at " << filename << std::endl;
-		return -1;
-	}
-
-	// Before filtering
-	imshow("Original image", image);
-	waitKey(0);
-
-	// Initialize filter. Kernel size 5x5, threshold 20
-	Ptr<epf::edgepreservingFilter> filter =
-	    epf::edgepreservingFilter::create(&image, &res, 5, 20);
-
-	// After filtering
-	imshow("Filtered image", res);
-	waitKey(0);
-
-	return 0;
-}
